Moved sat-mediator to its own directory
