+++ Dockerfile.sat
@@ -11,8 +11,9 @@ COPY packages/ ./packages/
 # Install dependencies
 RUN npm ci
 
-# Copy app
-COPY src/*.js ./
+COPY mediators/sat ./sat/
+WORKDIR /app/sat
+RUN npm ci
 
 # Run...
-CMD ["node", "sat-mediator.js"]
+CMD ["node", "src/index.js"]
