+++ mediators/sat/README.md
@@ -1,0 +1,2 @@
+# MDIP Satoshi mediator
+
