+++ mediators/sat/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "hyperswarm-mediator",
+  "version": "1.0.0",
+  "type": "module",
+  "description": "MDIP hyperswarm network mediator",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "David McFadzean <davidmc@gmail.com>",
+  "license": "MIT",
+  "dependencies": {
+    "bitcoin-core": "^4.2.0",
+    "dotenv": "^16.4.5",
+    "express": "^4.21.0",
+    "morgan": "^1.10.0"
+  }
+}
